Update 2.2
Added OrbWalking to have more control of your hero.
Fixed some early attacks with another stage of checking if it is gonna be an early attack. If so it will keep canceling attack until it calculates that it is not going to be an early attack anymore.
Added Weaver's Geminate Attack Damage calcualtion.
Script now prefer Lasthits rather than Denies.